requirements.txt
PyYAML==5.4.1
pdfservices-sdk==4.1.0
streamlit
PyMuPDF
tqdm
requests
pymongo
openai
python-dotenv
tiktoken